+++ demos/session_service/src/main/java/androidx/media3/demo/session/DemoPlaybackService.kt
@@ -29,6 +29,7 @@ import androidx.core.app.NotificationManagerCompat
 import androidx.datastore.core.DataStore
 import androidx.datastore.core.Serializer
 import androidx.datastore.dataStore
+import androidx.lifecycle.lifecycleScope
 import androidx.media3.cast.CastPlayer
 import androidx.media3.common.AudioAttributes
 import androidx.media3.common.Player
@@ -152,8 +153,7 @@ open class DemoPlaybackService : MediaLibraryService() {
   @OptIn(UnstableApi::class) // Player.listen
   private fun initializeSessionAndPlayer() {
     val player = buildPlayer()
-    // TODO: b/466098427 - Use scope tied to the lifecycle of DemoPlaybackService instead.
-    CoroutineScope(Dispatchers.Unconfined).launch {
+    lifecycleScope.launch {
       player.listenTo(Player.EVENT_IS_PLAYING_CHANGED, Player.EVENT_MEDIA_ITEM_TRANSITION) {
         storeCurrentMediaItem()
       }
@@ -165,8 +165,7 @@ open class DemoPlaybackService : MediaLibraryService() {
         .build()
 
     mediaLibrarySession.setCustomShuffleModeButton()
-    // TODO: b/466098427 - Use scope tied to the lifecycle of DemoPlaybackService instead.
-    CoroutineScope(Dispatchers.Unconfined).launch {
+    lifecycleScope.launch {
       player.listenTo(Player.EVENT_SHUFFLE_MODE_ENABLED_CHANGED) {
         mediaLibrarySession.setCustomShuffleModeButton()
       }